+++ CHANGELOG.md
@@ -12,7 +12,7 @@ All notable changes to this project will be documented in this file. This projec
 - Support for periodic execution of software timers. All overloads of `SoftwareTimer::start()` were extended with
 optional second argument with period. If the period is 0 (which is the default value), software timer's function is
 executed only once. Otherwise it will be executed periodically until the software timer is stopped.
-- `internal::UartBase` interface class, which can be used as a private base for devices using UART interface.
+- `devices::UartBase` interface class, which can be used as a private base for devices using UART interface.
 - `internal::UartLowLevel` interface class, which is a low-level interface to hardware-dependent UART driver.
 - `chip::UartLowLevel` class - interrupt-driven implementation of `internal::UartLowLevel` interface for *STM32F4*.
 - "Peripherals configuration" *Kconfig* menu for *STM32F4*, where low-level drivers for U[S]ART can be selected.